Refactor intent subtree into constant
There was some repetition of the string "/intent/%s" because intent.WatchPods
takes a raw string for its path. We don't want to take that string parameter
away because it makes the function more versatile, but we also want to ensure
that our code agrees on where the intent store keys actually live.

The easiest solution is to move the intent store's path prefix into a constant
and use that wherever we use WatchPods. In general, we know that WatchPods
without this constant is suspicious, but we can choose to use it that way if
necessary.