Add RegisterPodService to intent store
This function registers a consul service for a pod, along with the associated
health check. Because consul services are tied to a single agent in the cluster,
you have to run this function against the same agent that is running the pod
itself, ie the function has to be invoked by the preparer.