Remove leading slash on intent path
Turns out the consul API package is sensitive to this leading slash. It
gets concatenated to the API path prefix, which results in a double-slash
in the path. Consul 301s back to the canonical path when it sees the double
slash, and the API package errors out on any non-200 response code.

Here is the incriminating line: https://github.com/armon/consul-api/blob/master/kv.go#L171