Refactor shared arguments of preparer into struct
The three functions in orchestrate.go form a simple call chain. The
arguments at the top of the chain are threaded down to the functions
below, but the list is getting a bit long. We can refactor all the
arguments into a struct to provide the shared context between all
the functions.