Fix preparer tests
The preparer tests are for installAndLaunchPod, which doesn't actually use
most of the fields in the Preparer struct. So we don't need to change much.

I also added a comment for a sneaky aspect of installAndLaunchPod: it takes
its own sublogger as an argument. It does not use the Preparer's Logger, so
don't combine them.