Allow p2-launch to specify user
The user that a pod executes under can now be configured, so the consul
pod is not hardcoded to run as root. Unfortunately, this means you can't
p2-launch the consul pod - it tries to run consul under the "intent"
user.

You have to be root to run p2-launch anyway, so we can add a flag to
specify the user of the pod being launched to fix this issue.