Remove ppkv from p2-bootstrap
There are two packages in p2 that manipulate consul. The first is
armon/consul-api (a deprecated older version of hashicorp/consul/api).
When manipulating the key-value store with this package, the values
are treated as raw strings. Consul doesn't care what they are or what is
in them.

The second is square/p2/pkg/kv-consul. This is legacy code that wraps
around consul-api... but it assumes all the values in the key-value
store are JSON, so it encodes before writes and decodes after reads.
Unfortunately for p2, we were mixing the use of these packages in some
very critical places.

Because p2-bootstrap was using ppkv, the preparer's manifest was added
to the intent store as JSON-encoded YAML. To read this, you have to
decode the JSON first, before attempting a PodManifestFromBytes. On the
other hand, all the other manifests that went into the intent store were
sent there using intent.SetPod, which does not encode the manifest as
JSON.

This creates an unpleasant conundrum for intent.WatchPods. Some of the
manifests it reads will be JSON-encoded YAML, but others are just YAML.
To fix this, we had a ghetto JSON decoder, which just unescaped some key
characters (double quotes and newlines) if it saw a double quote at the
start of the string. This JSON "decoder" would correctly decode YAML
that had been encoded as a JSON string, but it would skip over YAML
values without raising any errors. None of us really knew why it was
working, but it worked.

Enter p2-inspect. One of p2-inspect's jobs is to read the intent store.
To do this, we added a new method to the intent store, intent.Pod
(analogous to reality.Pod). This was implemented using consul-api, so it
would fail every time it tried to read the preparer manifest. It would
try to decode the YAML, but the preparer manifest was actually JSON-
encoded YAML. We went around in circles trying to figure out where the
JSON encoding was actually happening, until eventually we found it
lurking in p2-bootstrap.

The ultimate fix is to remove ppkv from p2-bootstrap so that every
manifest is stored in consul as raw YAML.