Release file descriptors during tar unpacking
Deferring all the close() calls to the end results in a massive number
of open file descriptors. If you open a tarball with a lot of files in
it, there will eventually be too many.

Close() is an idempotent call, so we can close the file descriptor more
quickly to avoid this accumulation.