Activate launchables before generating runit services
This change is required to match the behavior of our legacy deployment
system. Some launchables use a post-activate script to generate their
launch scripts. If you try to generate the runit services before the
post-activate has run, there will be no launch scripts and therefore no
runit services. Because the runit service directories don't exist, an
error will occur when the preparer tries to invoke sv in the service
directory (it tries to cd there and it doesn't exist).