Eagerly create directories during tar unpack
This fixes a bug where a tarball contains folders of folders of files.
If the parent folder has no files in it, then it will never be chowned,
because it was created as part of makeAndChown. This function called
MkdirAll to create the directory chain, but then it only chowned the
last directory in that chain.