Test hooks in the integration test
This commit replaces the use of user.CreateUser with a
simple create_user shell script that is launched as a before_install
hook pod. This has the effect of verifying that the hook is
found in the intent store, installed correctly, and launched at
the correct time during the install / launch.

This commit includes a number of small fixes and changes to facilitate
that goal.