Augment manifests to support optional PGP signatures
With this commit, a clearsigned PGP message containing a manifest can
now be transparently parsed as if it were any other manifest. The
signature data will be preserved in the parsed manifest. Of course,
mutating this manifest in any way will invalidate the signature, but we
can't really show that without dramatic API modifications.