Add ACL token support to preparer
The preparer takes its ACL token from a file. The path to the file is
set in its CONFIG_PATH.

In the long run, ACL tokens are secrets, so they can't be stored in a
pod manifest (that's insecure). Instead, we can eventually move the
token file into a more secure secrets provisioning system (Square has a
system for this in-house). This is just convenient until we integrate
those systems into p2.