Fix tar hardlinks
Symlinks in a tarball are stored the way you would expect (the path of
the symlink is relative to the symlink itself), but hardlinks in a
tarball are handled differently (the path is relative to the tar root).
We need to resolve a relative path between the hardlink and its target
before we can create the link itself.