Notes on sv stop
Because stop and start are separated, the preparer must not terminate
when it tries to sv stop itself. It has to continue execution so that it
tries to sv restart itself later. This ensures that, when the preparer
does actually exit, runit will try to start it again.

Because of this, we cannot stop the launch if the preparer fails to stop
(because the preparer would not attempt to restart itself), and we
cannot pass the -w flag when trying to sv stop the preparer (otherwise
we will try to SIGKILL ourselves).