Clean up brief ACL timing issue
When a Consul server cluster converges on a leader for the first time,
it immediately begins raft replication between servers. But if you talk
to a server in the exact instant that the leader is elected, the ACL
data will not be replicated yet, and you may get a 403 ACL not found. A
brief sleep can push this error aside in this particular case.