Do not use entropy in tests
When we first implemented manifest signing, we started getting test
failures because those manifests were unsigned. So we started generating
keys in tests and signing them. Then we started getting failures because
Travis VMs didn't have enough entropy to generate keys over and over
again. So we started generating one keypair and caching it for the
entire test.

Then we got a failure because a Travis VM didn't have enough entropy to
generate the very first keypair. At this point, the best solution is to
not consume random bits at all. Generating primes will take a nontrivial
time regardless of the RNG source, so instead we hardcode the signing
entity for test purposes.