Do not return errors directly from consul API
Errors returned from the consul API may include URLs, which may include
ACL tokens. We can't afford to log those. Fishing through the token for
a URL field is only possible via reflection, which is expensive when it
would be done on every operation. Let's just encapsulate the errors and
return a less informative error instead.