Refactor authorization into its own module
The _authorization policy_ of the p2-preparer is the method by which it decides
whether it is safe to install and run a pod or launchable artifact on its node.
The existing preparer defined its policy in many different locations throughout
the code base, making it difficult to change the policy, and it wasn't entirely
clear what the policy was.

This commit refactors all authorization policy logic into a single module to
fix the above two problems. Different policies can be selected when the
Preparer object is constructed (e.g., via configuration or from unit tests)
without requiring internal changes. By grouping the logic together, it is
easier to understand the exact conditions that permit a pod to run.

Two policies are defined so far:
- `NullPolicy`: accepts everything. Used if the preparer's config does not
  specify a keychain file.
- `FixedKeyringPolicy`: Authorization is defined by a keyring. Any signature
  made by any key in the keyring authorizes any package (except "p2-preparer",
  which has an extra list of valid keys). The keyring is defined once at
  startup and never changed.