Dynamically load PGP keys when verifying signatures
The current implementation loads a keyring from a file once at startup. If the
keyring file changes after that, the preparer would need to be restarted to
pick up the changed file.  This commit adds the `FileKeyringPolicy`
authorization policy that will reload the file's contents if it has changed
since the last request for the keyring.