Implement consul session locking
Sessions can be used to acquire locks on keys in the kv store. Because
acquiring the lock involves writing to it, we can't lock on the keys of
the intent store directly (otherwise we might overwrite a meaningful
value), but we can create a separate key tree specifically for locking.