Refactor URI fetcher to support streaming
The "uri" module presented a nice interface for copying a file from an
arbitrary URI to a local file. But what if you just want to consume the data
stream and not store it? This commit refactors the `uri.Copier` type into the
`uri.Fetcher` interface, which exposes the URI data stream directly (you can
still make a local copy if you want).

By accessing the data stream, users can avoid allocating a temporary directory,
writing a file to it, operating on the file, then deleting it.