Flexible authorization configuration
Changes the layout of the preparer configuration configuration file to allow
for flexible construction of authorization policies. The configuration file
show how contain an "auth:" section which should contain a map. One required
field is "type", which specifies what authorization type the rest of that
section should be parsed as. Other fields in the "auth" section are defined by
the selected type.

The following types are currently defined:

none:  No authorization checks are performed. Handled by the `NullPolicy` type.
       No additional configuration needed.

keyring: Uses a keyring file to authorize deploys. Handled by the
         `FileKeyringPolicy` type. Additional configuration fields:
    - keyring: specifies the keyring file to use
    - authorized_deployers (optional): a list of full PGP key IDs that can
      deploy the preparer.