Refactor file-reloading logic out of FileKeyringPolicy
Watching a file for changes and reloading it is a behavior that is independent
of the particular file type (keyring) and use of the file (policy). This commit
separates these two components into distinct types. The `util.FileWatcher` type
contains logic for reloading from a file. Its logic was pulled out of
`FileKeyringPolicy`, which just specifies how to parse the keyring file and
what to do with it.