Parse URIs for fetching
Before this change, schemeless URIs and those with the file:// scheme
were not being parsed; instead the scheme was manually removed and the
rest of the URI was treated as the path. This resulted in a bug where
a file:// URI with an absolute path had its leading slash removed, eg
"file:///path/to/file" was parsed as "path/to/file".

For consistency, this change uses a parser to extract the scheme and
path instead. This causes query metacharacters to also be parsed, so
they have to be percent-escaped. For example, the path "/my?file" would
have to be written as "file:///my%3Ffile".