Authorize a pod based on its manifest's user, not its name
Changes the UserPolicy authorization policy to look for authorization based on
the user the pod will run as, not the name of the pod. This makes the tool more
similar to "sudo", in that users are authorized to run commands (deploy
software) as other users.

This commit also renames the configuration parameter "forbidden_pod_ids" to
"forbidden_users".  Even before this commit, it was blocking users in the
"run_as" field, but its name did not convey this.