Remove "forbidden_users" preparer option
The "forbidden_users" option was an extra authorization check performed after
the normal authorization policy allowed an app to be installed. It allowed
certain "run_as" usernames to be blacklisted. Since the "user" policy is
already verifying the "run_as" user, the extra check is redundant.

This commit removes the "forbidden_users" option from the config struct and all
internal uses of it. It can continue to be specified in the preparer's config
without error; it will simply have no effect.