Adds watch package
The watch package exposes the WatchHealth method. This
method monitors the reality tree in consul to determine
what services should be running on a node. WatchHealth
launches a health-checking go routine for each pod on
the node. The package also exposes the MonitorHealth method
which actually performs the health check on a particular
service.

This package will be used by the p2-preparer to enable
p2 to perform health checks without consul agents.