Preparer now gets consul token from file
preparerConfigs contain a ConsulTokenPath field.
Now that the preparer is talking to consul it
actually needs to use a consul token. There was
an unexposed method in the preparer package that
got the token string from the file path. This
commit exposes that method and the preparer now
uses LoadConsulToken to get the token so
WatchHealth can actually talk to consul.