fix integration test for health check
Had to fix logging in the watch package. Was trying
to reference nil pointers because PodWatch's logger
field was nil.

The manifest for hello which is populated in check.go
did not have the StatusPort or StatusHTTP fields set.

Integration test for watch pkg now asserts that all
services being health checked are passing. I don't
know why I didn't do that before...