Add "--show-consul-errors-unsafe" flag to show Consul errors
We don't want to show Consul errors all the time--they can leak ACL tokens--but
when debugging a problem involving Consul, that kind of information is very
important to understanding what's going on. This commit adds a command line
flag that will enable detailed log messages, to be used only when needed.