Fix failing kv unit test
The unit tests in .../pkg/pk/kv_test.go use Consul's testutil package to start
a live Consul server node for the duration of the test. To allow multiple tests
to run in parallel, each test server listens on a unique port for HTTP request.

This commit forces the unit test to use whatever port was actually allocated
for Consul instead of the default 8500.