Add feature flag for Consul health checks
To make it easier to experiement with Consul health checks disabled, this
commit adds a feature flag to the preparer's config, `no_consul_health`, that
will cause the preparer to quit registering services with the Consul agent if
true. It's expressed as a negated feature flag to ensure that existing
configuration (that doesn't include this key) defaults to false, i.e., _do_
use Consul health checks.