find*Result in health package handles empty slice inputs
In some cases findWorstResult and findBestResult were being
passed empty result slices. This was cause panics. Now when
they are passed an empty slice they return a HealthEmptyError.
HealthEmptyError wraps an error and allows callers to handle
it as they see fit.

Updates health_test to reflect these changes and includes a
test for empty slices passed to the aformentioned methods.