Cleanup health checks when shutting down the main monitor
The goroutine running `MonitorPodHealth` spawns other goroutines that handle
health checks for individual apps. This commit adds cleanup code to ensure that
the those other goroutines are shut down when the MonitorPodHealth goroutine is
itself shut down.

This is probably a moot point, since shutdown only occurs when the process is
in the process of exiting, but I'm adding it for completeness.