Fixes preparer config in integration tests
The write_kv_health key was added to the preparer config
struct. However when this key was added no changes were
made in the integration test preparer config. Not having
the write_kv_health field set caused the integration tests
to fail. Setting the field to true causes the test to pass.