Add StringValuePublisher
StringValuePublisher is a new type that implements a broadcast channel for a
string value that changes over time. The publisher contains some current value,
and that value is updated via messages on a channel. The publisher's value can
be subscribed to, which sends its current and future values over a channel.