Add session-based health checks
This commit adds an implementation of kp.HealthManager and kp.HealthUpdater
that save health status that are locked to a common session. Status are written
once with a high expiration time and a short-lived Consul session ensures that
if the preparer dies, the status will be removed.

The existing write-every-15-seconds implementation has been preserved, and is
still the default. A subsequent commit will enable the new method.