Increase unit test timeout to 20s
Unit tests that invoke Consul's testlib are slow to start, and now the "kp"
package has a lot of those tests. Worse, when the timeout is reached and tests
are run with "go test ./...", the test process just seems to hang. This commit
ups the timeout to 20s to make it much more likely that unit tests will finish on time.

PR #223 includes a more thorough handling of this problem. We'll iron out the
differences once one of these PRs is committed.