Only update KV health for status changes
Some services include unique output in their status checks, e.g., a current
timestamp, and this was causing the services' KV health checks to be updated on
every check. With this commit, a new health check will be written only when
"Status" changes in a check result. "Output" is still present, but it cannot be
relied upon to be the most recent value.