Label storage implementation
Started by @anthonybishopric. This code provides a generalized interface
for storing labels (maps of strings to strings) on "things" (namely pods
and nodes), and a consul-backed implementation of that interface.

The implementation stores labels as JSON objects under the path
/labels/<type>/<id>. To safely mutate the object, CAS operations are
necessary, so the implementation supports automatic retries for CAS
failure. Tests included.