Global hook installs
This commit adds global hooks, which fire on all hook events. The event
in question is indicated by an environment variable passed to the hook,
which was implemented as part of the BEFORE_UNINSTALL hook feature. In
the long run, the goal is to eliminate typed hooks altogether. This
should make implementation easier for hooks that care about multiple
events, such as a hook with paired install and uninstall behaviors.