Enable/disable errors should not flip success
enable and disable are fire-and-forget scripts. It's important to log
their output if they fail, but they should not cause the success return
flag of a launch/halt to become false. To distinguish between the error
sources, some types have been added.

Motivation for this change: enable failures cause the recently added
preparer.resolvePair function to return false. As a result, the working
flag in handlePods is not flipped, which causes resolvePair to be called
again with the same ManifestPair, and to keep returning false. This spin
problem eventually stops itself when a new intent set is returned, which
updates the ManifestPair to contain the new reality manifest as well.

The problem was never observed before because reality manifests used to
be read as part of preparer.installAndLaunchPod. Therefore, if the
enable failed, then the handlePods loop would spin, but because
installAndLaunchPod read the new reality manifest, it would not make any
further changes or log messages.