Add params package
Adds a new "params" package which provides a less-formal way to specify
configuration parameters.  Each parameter can be declared in the package where
it's needed, and a single call to `params.Parse()` will set all defined
parameters.

This is mainly intented to permit some configurability for values that would
otherwise be declared as a package constant, or where there are too many layers
between a program's main configuration parser and the use of that configuration
value.