Fix ZipResultSets conditions
If the intent list was exhausted before the reality list, then iID would
be the empty string, and rID would be non-empty. Therefore, iID < rID,
and the code would try to retrieve the last intent manifest (which would
promptly panic).

This would never happen if reality was exhausted, since rID == "" would
be caught by the first case before reaching the second case (which would
have panicked, if it had been executed).