Fix bugs in callers of WatchPods()
Recently the behavior of WatchPods() changed such that it can return an
empty []ManifestResult for a pod if the key does not exist in the kv
store. pkg/replication had a place where it would not handle this event
properly. If the key being watched doesn't exist, it should block until
it does and matches the sha of the manifest being replicated