Store a replication controller in a single key
We were storing a replication controller as a tree of keys under its id,
one for each field. This approach makes it very easy to mutate fields of
the RC. However, it has a nasty side effect: adding a new RC key-by-key
will trigger watchers before the RC is complete. As a result, the watch
will read a partial RC with some of its fields missing.

To avoid this partial-read issue, we use the same approach as the consul
label applicator - the entire RC is a single JSON blob that uses CAS
operations to safely update individual fields.

Unmarshaling an RC from JSON presents its own problems, since two of the
fields are interfaces, and unmarshaling a non-JSON-null into a nil
interface is impossible. We have to provide implementations of the json
Marshaler/Unmarshaler interfaces for this purpose.