Use unmanaged locks for RCs
As part of a longer-term goal, it would be good to take the session
management out of kp.Lock, so that we can use it with the session
manager. But for now that refactor is too large. Adding this new
unmanaged constructor lets us use Lock and ConsulSessionManager for now,
without mandating a larger change.

Since sessions are no longer part of most log lines, the logging in the
rc farm loop is also cleaned up using a sublogger.