Fixes to pkg/runit
Previously the runit package did not account for when services were
in a down state. This would have resulted in pkg/runit returning an
error instead of stat information.

Exposes the child status field, which is necessary for the runit
monitor to accurately ascertain if the service is up and healthy.