Count only real nodes as healthy in updates
An RU can potentially count a node as healthy once the RC has scheduled
it, but the preparer on that node might not have finished the update. As
a result, the node might be healthy because it hasn't actually finished
restarting and rolling over to the new version. This tricks the RU into
progressing, when it should actually block for that node to finish and
then become healthy.