Use "cgroup.procs" for cgroup membership
This commit changes p2-exec to use a different method for changing its cgroup
membership: it will write "0" to the cgroup's "cgroup.procs" file. Writing to
this file will change membership for its thread group (e.g., the entire
process) instead of an individual thread. This may be a fix for a bug in which
a launched service is observed to be running in the root cgroup.