Add RestartPolicy to pod manifest.
If unspecified, the RestartPolicy defaults to "always". P2 has
previously only supported processes that restart after exit. We now want
to support deploying a runit service that is meant to be run only once.
If RestartPolicy is set to "never", the service will be started with
`sv once` which will result in the service being left down upon exit.