Allow cleaning up old RC to be disabled
Example use case: you create an RU to deploy a single node of a new
version, as a canary. Then, later, you want to complete the rollout and
go to all the nodes. Rather than implementing a way for the original RU
to "indefinitely pause", we can split this flow into two RUs, and tell
the first RU to not delete the old RC, so the second one can use it.