Fix double-launch bug
Adds a quick fix for a bug in which app will be installed twice. More details
are in the patch itself.

Tested manually by watching scheduling a manifest, waiting for the preparer to
pick it up, then scheduling the same thing again. This procedure is sufficient
to get the earlier version to install twice. With this change, it doesn't.