Allow shorthand for memory limits.
This change allows manifest authors to specify a shorthand for
the byte limits of their cgroups. Previously, all `memory` entries
for the `cgroup` stanza had to be in the form of an integer
number of bytes. Authors may now use a shorthand as defined by
the package `github.com/square/p2/pkg/util/size`.

An author wishing to limit their launchable to a maximum resident
memory limit of 4 Gibibytes may use the shorthand `4G` instead of
`4294967296`.

The old form of specifying an integer limit of bytes is also supported.