Preserve cgroup serialization
Changes the cgroup.Config struct so that it preserves its old serialization.
For example, an input of "4K" is valid and parses to 4096, but regardless of
the input format, it will always serialize as 4096. This will allow older tools
to continue to work with the update input format, and it will ensure that there
is a single canonical representation of any manifest.