Do not return partial lists from WatchPods
If a parse error is encountered partway through the key list, we should
discard the entire list and error out. This behavior ensures that, if a
list is returned, that it is complete, ie contains all the keys that
were requested.

In particular, if there's a bad manifest in /intent, and we return the
partial list, the pod corresponding to that bad manifest may be deleted
by the preparer because it's not in the list. Ideally the preparer loop
should be refactored to use NewKeyWatcher on both intent and reality
together, but until then, this change at least hardens it against
unintended deletions.