Fix ordering violation in stream package
A bug in the stream package was found while running unit tests for an unrelated
change. The test failure was:

    --- FAIL: TestUnsubscribeIndependence (0.00s)
            publisher_test.go:254: sub2 got unexpected sequence: [foo]

A subscriber should have received the current value of the StringValuePublisher
as of the time of the Subscribe() call; in this execution, that initial value
was skipped.

Root cause: when the subscriber unsubscribes, it closes its "canceled" channel
to signal the publisher that it shouldn't wait when sending to the subscriber.
Then, in the presence of an active reader, it is possible for sendFirst() to
non-deterministically choose to abort its send while the main publisher read()
loop non-deterministically chooses to send a subsequent value.

This commit fixes the ordering violation by ensuring that sendFirst() always
finishes executing before read() can publish a value, then by checking if the
subscription has been canceled before attempting to send a new value.

Tested with ~300,000 executions of the unit test without any failures.