Add WithSession() helper
Adds the WithSession() helper method for working with Consul sessions. Launches
a function in a goroutine when a session appears; asks the function to exit
when the session dies. This method provides a simple model for processes that
require sessions but don't want the added complexity of handling session
refreshes.