Refactor session locking out of RC/RU farms
Refactors the farm main loops to avoid managing sessions themselves. Deferring
to WhileSession(), the main loops will be started when a session exists, and it
will be stopped when the session is lost. Farms no longer need to handle the
complexities of changing sessions, and their code can be more focused on
changing data and its handler routines.