Wait for hooks to sync at least once before managing pods
If a preparer is restarted and simultaneous updates to hooks and one or
more pods are pending, there is a race where the preparer might process
the pod before processing the hooks. This can leave the pod in an
unexpected state with no mechanism for repair.

This PR makes the loop that watches pods wait until hooks are synced at
least once before beginning