launch: Rename ID to service ID
We are about to commit a change that requires launchables to know their
own ID. Currently, a launchable ID is the concatenation of the pod ID
and the launchable ID specified in the launchable stanza.
For example, with pod foo launchable bar, the launchable ID would be
`foo__bar`.

We will have use cases that require us to just have the launchable ID,
`bar`. So we need to store just the launchable ID at some point. To
avoid confusion with existing uses of service ID (which are perfectly
valid because they needed to be used for runit, cgroups, etc.), we
rename the existing ID to ServiceID.

A future commit will then add a new ID field that just has the
launchable ID without the pod ID prepended.

To verify that all existing uses have been renamed, one should be able
to build the repo at this commit.