launch: Add launchable IDs
The preceding commit "Rename ID to service ID" gave us the ServiceID
field that has the concatenation of the pod ID and the launchable ID
specified in the launchable stanza (for example `foo__bar` for pod foo
launchable bar).

Now we want to have a field with just the launchable ID such as `bar`.
It will just be called ID.

The next commit will put this to use.