Make launchable ID in test manifests consistent
We are going to deprecate launchable_id at some point, isntead just
taking it from the key in the manifest. When this happens, we'll want to
make sure the two are the same. We should bring our own manifests in
compliance with this policy.

Another effect of this change is that it lets us actually test whether
we distinguish the pod and launchable IDs correctly. If we had mixed up
the ordering of the concatenation of the service ID, we might have
gotten a value of `app__hello` instead of `hello__app`. This error would
have been undetectable had the launchable ID also been "hello".