pod: Write LAUNCHABLE_ID to launchable env dir
For example, in a pod named "foo" with a launchable named "bar", we
would write "bar".

This allows launchables to easily introspect the ID they were given in
their pod manifest.
It is useful to allow launchables to know this instead of forcing them
to assume they have been given a particular ID.
We already expose LAUNCHABLE_ROOT (and theoretically we could say
launchables have to extract their ID from the ROOT) but it's easier for
the launchables if we just tell them their ID.