Make manifest.GetConfig() return a deep copy of the config map
This avoids modification of config maps returned by GetConfig() from
mutating the underlying manifest or eachother.

Copying is done by marshaling to yaml and then unmarshaling into a new
map. This introduces a chance for an error, but by doing error checking
when configs are set we can guarantee that in practice these errors
should not occur, and if they do then something is very wrong and thus
we panic.