bin/p2-log-bridge must have a blocking STDIN
Comments inline, tl;dr: svlogd sets the named pipe it uses as its stdin
(recall that stdout, stderr of the program under log are writers to this
pipe) to be a non-blocking pipe.  This is not OK for go, as go's File
struct does not handle the EAGAIN messages that are common to see from
the read(2) syscall when reading a non-blocking pipe.